generate date for readme each time cause to much noise with on git

+++ docsible/cli.py
@@ -15,7 +15,7 @@ DOCSIBLE_END_TAG = "<!-- DOCSIBLE END -->"
 timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
 
 def get_version():
-    return "0.6.6"
+    return "0.6.7"
 
 def manage_docsible_file_keys(docsible_path):
     default_data = {
@@ -37,13 +37,13 @@ def manage_docsible_file_keys(docsible_path):
         with open(docsible_path, 'r') as f:
             existing_data = yaml.safe_load(f) or {}
         
-        # Update dt_update field
-        existing_data['dt_update'] = datetime.now().strftime('%d/%m/%Y')
-        
         updated_data = {**default_data, **existing_data}
-        with open(docsible_path, 'w', encoding='utf-8') as f:
-            yaml.dump(updated_data, f, default_flow_style=False)
-        print(f"Updated {docsible_path} with new keys.")
+        if updated_data != existing_data:
+            # Update dt_update field if docsible keys added
+            existing_data['dt_update'] = datetime.now().strftime('%d/%m/%Y')
+            with open(docsible_path, 'w', encoding='utf-8') as f:
+                yaml.dump(updated_data, f, default_flow_style=False)
+            print(f"Updated {docsible_path} with new keys.")
     else:
         with open(docsible_path, 'w', encoding='utf-8') as f:
             yaml.dump(default_data, f, default_flow_style=False)
